+++ django_cat/agent/models.py
@@ -1,3 +1,4 @@
+from typing import Any, Union
 from django.db import models
 from django.dispatch import receiver
 from cheshire_cat.types import AgentRequest, Agent as AgentModel
@@ -8,6 +9,11 @@ from app.signals import server_start
 from icecream import ic
 
 
+class AgentManager(models.Manager):
+    def filter(self, *args, **kwargs):
+        """Never return the default agent"""
+        return super().filter(*args, **kwargs).exclude(agent_id="default")
+
 class Agent(BaseUserModel):
     agent_id = models.CharField(max_length=255, null=True, blank=True, default=None)
     name = models.CharField(max_length=255, default="")
@@ -17,23 +23,30 @@ class Agent(BaseUserModel):
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
 
+    objects = AgentManager()
+
     class Meta:
         ordering = ("user", "-updated_at")
 
-    def model_dump(self) -> AgentRequest:
+
+    @property
+    def agent(self) -> Union[AgentModel, AgentRequest]:
         if self.agent_id is None:
             return AgentRequest(
                 name=self.name,
                 instructions=self.instructions,
                 metadata=self.metadata
-            ).model_dump()
+            )
         else:
             return AgentModel(
                 id=self.agent_id,
                 name=self.name,
                 instructions=self.instructions,
                 metadata=self.metadata
-            ).model_dump()
+            )
+        
+    def model_dump(self) -> dict[str, Any]:
+        return self.agent.model_dump()
         
     def full_model_dump(self) -> AgentModel:
         return AgentModel(
@@ -44,14 +57,19 @@ class Agent(BaseUserModel):
         ).model_dump()
     
     @staticmethod
-    def get_default(user=None):
-        return Agent.objects.get_or_create(agent_id="default", user=user)[0]
+    def get_default():
+        from users.models import UserProfile
+        return Agent.objects.get_or_create(agent_id="default", user=UserProfile.get_admin().user)[0]
     
     @property
     def is_default(self):
         return self.agent_id == "default"
     
     def create_agent(self, save=True):
+        # If is default, is not necessary to create an agent
+        if self.is_default:
+            return self.agent
+        
         agent = self.client.create_agent(self)
         self.agent_id = agent.id
 
@@ -61,17 +79,24 @@ class Agent(BaseUserModel):
         return agent
 
     def save(self, *args, **kwargs):
+        if self.is_default:
+            self.name = "Default"
+
+        
         super().save(*args, **kwargs)
 
         if self.agent_id is None:
             self.create_agent()
-        elif self.pk:
+        elif self.pk and not self.is_default:
             self.client.update_agent(self)
 
     def delete(self, *args, **kwargs):
         self.client.delete_agent(self.agent_id)
         
         return super().delete(*args, **kwargs)
+    
+    def __str__(self):
+        return f"{self.name} ({self.agent_id})"
     
 @receiver(server_start)
 def create_agents_on_server_start(sender, **kwargs):